PenStrokeOrderJudgement/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
brotli=1.0.9=hb7f2c08_8
brotli-bin=1.0.9=hb7f2c08_8
ca-certificates=2022.12.7=h033912b_0
certifi=2022.12.7=pypi_0
charset-normalizer=3.1.0=pypi_0
colorama=0.4.6=pypi_0
cycler=0.11.0=pyhd8ed1ab_0
cython=0.29.34=pypi_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h3f81eb7_1
idna=3.4=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=hb7f2c08_3
kiwisolver=1.4.4=py37h229a17a_0
lcms2=2.12=hf1fd2bf_0
lerc=4.0.0=hb486fe8_0
libblas=3.9.0=16_osx64_openblas
libbrotlicommon=1.0.9=hb7f2c08_8
libbrotlidec=1.0.9=hb7f2c08_8
libbrotlienc=1.0.9=hb7f2c08_8
libcblas=3.9.0=16_osx64_openblas
libcxx=16.0.2=hd57cbcb_0
libdeflate=1.14=hb7f2c08_0
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=11_3_0_h97931a8_31
libgfortran5=12.2.0=he409387_31
liblapack=3.9.0=16_osx64_openblas
libopenblas=0.3.21=openmp_h429af6e_3
libpng=1.6.39=ha978bb4_0
libsqlite=3.40.0=ha978bb4_1
libtiff=4.4.0=h6268bbc_5
libwebp-base=1.3.0=hb7f2c08_0
libxcb=1.13=h0d85af4_1004
libzlib=1.2.13=hfd90126_4
llvm-openmp=16.0.2=hff08bdf_0
matplotlib=3.5.3=py37hf985489_2
matplotlib-base=3.5.3=py37h3748cd6_2
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.3=h96cf925_1
numpy=1.21.6=py37h345d48f_0
opencv-python=4.7.0.72=pypi_0
openjpeg=2.5.0=h5d0d7b0_1
openssl=3.1.0=h8a1eda9_2
packaging=23.1=pyhd8ed1ab_0
pillow=9.5.0=pypi_0
pip=23.1.1=pyhd8ed1ab_0
protobuf=3.20.3=pypi_0
pthread-stubs=0.4=hc929b4f_1001
pyparsing=3.0.9=pyhd8ed1ab_0
python=3.7.12=hf3644f1_100_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=3_cp37m
pyyaml=6.0=pypi_0
readline=8.2=h9e318b2_1
requests=2.28.2=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
setuptools=67.7.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.40.0=h9ae0607_1
tensorboardx=2.6=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h5dbffcc_0
torch=1.13.1=pypi_0
torchaudio=0.13.1=pypi_0
torchvision=0.14.1=pypi_0
tornado=6.2=py37h994c40b_0
tqdm=4.65.0=pypi_0
typing-extensions=4.5.0=hd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
urllib3=1.26.15=pypi_0
wheel=0.40.0=pyhd8ed1ab_0
xorg-libxau=1.0.9=h35c211d_0
xorg-libxdmcp=1.1.3=h35c211d_0
xz=5.2.6=h775f41a_0
yacs=0.1.8=pypi_0
zstd=1.5.2=hbc0c0cd_6
